backend/requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
firebase-admin==6.0.0
celery==5.2.3
redis==4.3.4
python-escpos==3.0.0
gunicorn==20.1.0 